+++ add-on/gui/provider/ai_config_layout_openai.py
@@ -1,0 +1,48 @@
+from .ai_config_layout import AIConfigLayout
+
+class AIConfigLayoutOpenAI(AIConfigLayout):
+    def __init__(self, config_window, conf) -> None:
+        super().__init__(config_window, conf)
+
+    def advanced(self):
+        self.text("Advanced", bold=True)
+        self.text('<a href="https://platform.openai.com/docs/api-reference/chat/create#chat/create-temperature">More Details</a>', html=True, size=10)
+
+        number_input = self.number_input(
+            self.get_key("temperature"),
+            "Sampling temperature:",
+            decimal=True,
+            minimum=0,
+            maximum=2,
+            tooltip="Between 0 and 2. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.",
+        )
+        number_input.setValue(self.conf.get(self.get_key("temperature")))
+
+        number_input = self.number_input(
+            self.get_key("max_tokens"),
+            "Max Token:",
+            minimum=0,
+            maximum=2147483647,
+            tooltip="The maximum number of tokens to generate in the chat completion. The total length of input tokens and generated tokens is limited by the model's context length.",
+        )
+        number_input.setValue(self.conf.get(self.get_key("max_tokens")))
+
+        number_input = self.number_input(
+            self.get_key("presence_penalty"),
+            "Presence Penalty:",
+            decimal=True,
+            minimum=-2.0,
+            maximum=2.0,
+            tooltip="Number between -2.0 and 2.0. Positive values penalize new tokens based on whether they appear in the text so far, increasing the model's likelihood to talk about new topics.",
+        )
+        number_input.setValue(self.conf.get(self.get_key("presence_penalty")))
+
+        number_input = self.number_input(
+            self.get_key("frequency_penalty"),
+            "Frequency Penalty:",
+            decimal=True,
+            minimum=-2.0,
+            maximum=2.0,
+            tooltip="Number between -2.0 and 2.0. Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model's likelihood to repeat the same line verbatim.",
+        )
+        number_input.setValue(self.conf.get(self.get_key("frequency_penalty")))